minor: remove global component registration
