+++ apps/docs/nuxt.config.ts
@@ -34,7 +34,6 @@ export default defineNuxtConfig(
       components: [
         {
           path: '~/components',
-          global: true,
           pattern: '**/*.vue',
           ignore: ['**/examples/*.vue', '**/tests/*.vue'],
           pathPrefix: false,
