+++ apps/learn/nuxt.config.ts
@@ -26,7 +26,6 @@ export default defineNuxtConfig({
   components: [
     {
       path: '~/components',
-      global: true,
       pattern: '**/*.vue',
       ignore: ['**/examples/*.vue', '**/tests/*.vue'],
       pathPrefix: false,
