+++ packages/nuxt-core/layers/mdc/nuxt.config.ts
@@ -1,12 +1,14 @@
 export default defineNuxtConfig({
   mdc: {
     components: {
+      prose: false,
       map: {
         Icon: "MDCIcon",
         Container: "MDCContainer",
         ErrorBanner: "MDCErrorBanner",
         Button: "MDCButton",
         NetlogoCommand: "NetLogoCommand",
+        a: "ProseA",
       },
     },
   },
